Все переменные с пл. запятой переписаны на double

+++ src/TaskTwo.java
@@ -6,21 +6,21 @@ public class TaskTwo {
     public static void main(String[] args) {
         // reading eps
         Scanner consoleInput = new Scanner(System.in);
-        float eps;
+        double eps;
         do {
             System.out.print("eps > 0; eps = ");
-            eps = consoleInput.nextFloat();
+            eps = consoleInput.nextDouble();
         } while (eps <= 0);
 
         // calculation and output n
-        float bufn;
-        bufn = ((float)sqrt(eps) - eps) / eps;
+        double bufn;
+        bufn = (sqrt(eps) - eps) / eps;
         int n;
         n = (int)bufn + 1;
         System.out.printf("n = %d\n", n);
 
         // calculation and output ai
-        float ai;
+        double ai;
         for (int i = 1; i <= n; ++i) {
             ai = 1.0f / ((i + 1) * (i + 1));
             System.out.printf("a%d = %f\n", i, ai);
